Inquiry: "Inquired at" can be provided as a string
The population of the database is done via Alice fixtures
and so is the inquiredAt field in the invoices.
Unfortunately date as string will be forwarded as string
and not as DateTime object,
so using the value later will cause an error.
Therefor all incoming data will be transformed into a DateTime object
if they are scalar.

- \Crmp\AcquisitionBundle\Entity\Inquiry::setInquiredAt accepts strings now and turns them into date time.