UnitTests: Pre-Mock the controller
Mocks of the controller are used in almost any controller test cases from now on.
So far it takes a long to mock the controller which occur in almost every single test
and are therefor code duplicates.
For easier testing and lesser duplicates the mock will be created in every setup.

Note: This might lead to slower tests.
So override the ::setUp for specific cases
or undo the changes made in this commit.